refactor: reorganize components folder structure
